+++ stories/system/theme/tokens/semantic-token.stories.tsx
@@ -1,39 +1,7 @@
-import {
-  UIProvider,
-  extendTheme,
-  Box,
-  Semantics,
-  useColorScheme,
-  ThemeTokens,
-} from '@yamada-ui/react'
+import { UIProvider, extendTheme, Box, Semantics } from '@yamada-ui/react'
 
 export default {
   title: 'System / Theme / Tokens',
-}
-
-export const useColorSchemeToken = () => {
-  const { colorScheme } = useColorScheme()
-
-  const colors: ThemeTokens = {
-    primary: ['#2196F3', '#F44336'],
-  }
-
-  const theme = extendTheme({ colors })()
-
-  return (
-    <UIProvider theme={theme}>
-      <Box
-        bg='primary'
-        p='md'
-        rounded='md'
-        color='white'
-        transitionProperty='all'
-        transitionDuration='normal'
-      >
-        The current colorScheme is "{colorScheme}"
-      </Box>
-    </UIProvider>
-  )
 }
 
 export const useSemanticToken = () => {
